FromMinDump: fix _file_size field width

+++ elements/analysis/frommindump.cc
@@ -53,7 +53,7 @@ int FromMinDump::configure(Vector<String> &conf, ErrorHandler *errh) {
           .read_or_set("BURST", _burst, 32)
           .read_or_set("VERBOSE", _verbose, 0)
           .read_or_set("LIMIT", _limit, -1) // By default read all the file
-          .read_or_set("TIMES", _times,1) // For 1 time only. -1 is infinite loop
+          .read_or_set("TIMES", _times, 1) // For 1 time only. -1 is infinite loop
           .read_or_set("LOOP_CALL", loop_call, "")
           .read_or_set("DPDK", _dpdk, false)
 
@@ -106,10 +106,10 @@ int FromMinDump::initialize(ErrorHandler *errh) {
     return errh->error("Could not allocate memory to preload file!");
 
   fseek(_f, 0, SEEK_SET);
-  int ret = fread(_file_data, 1, _file_size, _f);
+  size_t ret = fread(_file_data, 1, _file_size, _f);
   if (ret != _file_size)
-    return errh->error("Error while reading the file in memory to %p (read %d "
-                       "bytes, expected %d)!",
+    return errh->error("Error while reading the file in memory to %p (read %lld "
+                       "bytes, expected %lld)!",
                        _file_data, ret, _file_size);
 
   fclose(_f);
